♻️🏢 Split route files
