+++ apps/admin/src/routes/_layout.show.exhibitors.$id.edit.situation/form.ts
@@ -4,8 +4,9 @@ import { createStrictContext } from "@animeaux/core";
 import { useForm as useFormBase } from "@conform-to/react";
 import { getZodConstraint, parseWithZod } from "@conform-to/zod";
 import { useFetcher, useLoaderData } from "@remix-run/react";
-import { ActionSchema } from "./action";
-import type { action, loader } from "./route";
+import { actionSchema } from "./action-schema";
+import type { action } from "./action.server";
+import type { loader } from "./loader.server";
 
 export function useFormRoot() {
   const { exhibitor } = useLoaderData<typeof loader>();
@@ -20,7 +21,7 @@ export function useFormRoot() {
 
   const [form, fields] = useFormBase({
     id: "exhibitor-situation",
-    constraint: getZodConstraint(ActionSchema),
+    constraint: getZodConstraint(actionSchema),
     shouldValidate: "onBlur",
     lastResult:
       fetcher.data != null && "submissionResult" in fetcher.data
@@ -34,7 +35,7 @@ export function useFormRoot() {
     },
 
     onValidate: ({ formData }) =>
-      parseWithZod(formData, { schema: ActionSchema }),
+      parseWithZod(formData, { schema: actionSchema }),
   });
 
   return [form, fields, fetcher] as const;
