+++ apps/admin/src/routes/_layout.show.exhibitors.$id.edit.situation/loader.server.ts
@@ -1,0 +1,31 @@
+import { db } from "#core/db.server";
+import { assertCurrentUserHasGroups } from "#current-user/groups.server";
+import { safeParseRouteParam } from "@animeaux/zod-utils";
+import { UserGroup } from "@prisma/client";
+import type { LoaderFunctionArgs } from "@remix-run/node";
+import { json } from "@remix-run/node";
+import { routeParamsSchema } from "./route-params";
+
+export async function loader({ request, params }: LoaderFunctionArgs) {
+  const currentUser = await db.currentUser.get(request, {
+    select: { groups: true },
+  });
+
+  assertCurrentUserHasGroups(currentUser, [
+    UserGroup.ADMIN,
+    UserGroup.SHOW_ORGANIZER,
+  ]);
+
+  const routeParams = safeParseRouteParam(routeParamsSchema, params);
+
+  const exhibitor = await db.show.exhibitor.findUnique(routeParams.id, {
+    select: {
+      name: true,
+      isVisible: true,
+      locationNumber: true,
+      standNumber: true,
+    },
+  });
+
+  return json({ exhibitor });
+}
