+++ apps/admin/src/routes/_layout.show.exhibitors.$id.edit.situation/route-params.ts
@@ -1,0 +1,5 @@
+import { zu } from "@animeaux/zod-utils";
+
+export const routeParamsSchema = zu.object({
+  id: zu.string().uuid(),
+});
