+++ apps/admin/src/routes/_layout.show.exhibitors.$id.edit.situation/route.tsx
@@ -1,50 +1,15 @@
 import { Action } from "#core/actions";
 import { ErrorPage, getErrorTitle } from "#core/data-display/error-page";
-import { db } from "#core/db.server";
 import { PageLayout } from "#core/layout/page";
-import { Routes } from "#core/navigation";
 import { getPageTitle } from "#core/page-title";
-import { assertCurrentUserHasGroups } from "#current-user/groups.server";
-import { safeParseRouteParam, zu } from "@animeaux/zod-utils";
-import type { SubmissionResult } from "@conform-to/react";
 import { getFormProps } from "@conform-to/react";
-import { parseWithZod } from "@conform-to/zod";
-import { UserGroup } from "@prisma/client";
-import type { ActionFunctionArgs, LoaderFunctionArgs } from "@remix-run/node";
-import { json } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
-import type { MergeExclusive } from "type-fest";
-import { ActionSchema } from "./action";
 import { FieldsetSituation } from "./fieldset-situation";
 import { FormProvider, useFormRoot } from "./form";
+import type { loader } from "./loader.server";
 
-export async function loader({ request, params }: LoaderFunctionArgs) {
-  const currentUser = await db.currentUser.get(request, {
-    select: { groups: true },
-  });
-
-  assertCurrentUserHasGroups(currentUser, [
-    UserGroup.ADMIN,
-    UserGroup.SHOW_ORGANIZER,
-  ]);
-
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const exhibitor = await db.show.exhibitor.findUnique(routeParams.id, {
-    select: {
-      name: true,
-      isVisible: true,
-      locationNumber: true,
-      standNumber: true,
-    },
-  });
-
-  return json({ exhibitor });
-}
-
-const RouteParamsSchema = zu.object({
-  id: zu.string().uuid(),
-});
+export { action } from "./action.server";
+export { loader } from "./loader.server";
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => {
   return [
@@ -57,45 +22,6 @@ export const meta: MetaFunction<typeof loader> = ({ data }) => {
     },
   ];
 };
-
-type ActionData = MergeExclusive<
-  { redirectTo: string },
-  { submissionResult: SubmissionResult<string[]> }
->;
-
-export async function action({ request, params }: ActionFunctionArgs) {
-  const currentUser = await db.currentUser.get(request, {
-    select: { groups: true },
-  });
-
-  assertCurrentUserHasGroups(currentUser, [
-    UserGroup.ADMIN,
-    UserGroup.SHOW_ORGANIZER,
-  ]);
-
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const formData = await request.formData();
-
-  const submission = parseWithZod(formData, { schema: ActionSchema });
-
-  if (submission.status !== "success") {
-    return json<ActionData>(
-      { submissionResult: submission.reply() },
-      { status: 400 },
-    );
-  }
-
-  await db.show.exhibitor.update(routeParams.id, {
-    isVisible: submission.value.isVisible,
-    locationNumber: submission.value.locationNumber || null,
-    standNumber: submission.value.standNumber || null,
-  });
-
-  return json<ActionData>({
-    redirectTo: Routes.show.exhibitors.id(routeParams.id).toString(),
-  });
-}
 
 export function ErrorBoundary() {
   return (
